client: rename `Bank Value` plugin to `Bank`
