+++ runelite-client/src/main/java/net/runelite/client/plugins/bank/BankConfig.java
@@ -23,14 +23,14 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package net.runelite.client.plugins.bankvalue;
+package net.runelite.client.plugins.bank;
 
 import net.runelite.client.config.Config;
 import net.runelite.client.config.ConfigGroup;
 import net.runelite.client.config.ConfigItem;
 
-@ConfigGroup("bankvalue")
-public interface BankValueConfig extends Config
+@ConfigGroup("bank")
+public interface BankConfig extends Config
 {
 	@ConfigItem(
 		keyName = "showGE",
